use xUnit features to set up db tests
