+++ src/TCode.r2rml4net.Tests/DatabaseSchemaReader/SqlServerCollection.cs
@@ -1,0 +1,9 @@
+﻿using Xunit;
+
+namespace TCode.r2rml4net.Tests.DatabaseSchemaReader
+{
+    [CollectionDefinition("SQL Server")]
+    public class SqlServerCollection : ICollectionFixture<SqlServerFixture>
+    {
+    }
+}
