+++ src/TCode.r2rml4net/RDB/DatabaseSchemaReader/DatabaseSchemaAdapter.cs
@@ -36,6 +36,7 @@
 // terms.
 #endregion
 using System.Collections.Generic;
+using System.Diagnostics;
 using System.Linq;
 using DatabaseSchemaReader;
 using DatabaseSchemaReader.DataSchema;
